+++ tfs-basic-domain/DSC/CreateDC.ps1
@@ -12,7 +12,7 @@ configuration CreateDC
         [Int]$RetryIntervalSec = 30
     ) 
     
-    Import-DscResource -ModuleName xActiveDirectory, xNetworking
+    Import-DscResource -ModuleName ActiveDirectoryDsc, xNetworking
     [System.Management.Automation.PSCredential] $DomainCreds = New-Object System.Management.Automation.PSCredential ("${DomainName}\$($Admincreds.UserName)", $Admincreds.Password)
     $Interface = Get-NetAdapter | Where-Object Name -Like "Ethernet*" | Select-Object -First 1
     $InterfaceAlias = $($Interface.Name)
@@ -70,11 +70,11 @@ configuration CreateDC
             DependsOn = "[WindowsFeature]ADDSInstall"
         }
          
-        xADDomain FirstDS 
+        ADDomain FirstDS 
         {
             DomainName                    = $DomainName
-            DomainAdministratorCredential = $DomainCreds
-            SafemodeAdministratorPassword = $DomainCreds
+            Credential                    = $DomainCreds
+            SafeModeAdministratorPassword = $DomainCreds
             DependsOn                     = @("[WindowsFeature]ADDSInstall")
         } 
     }